Unify tensor classes

+++ src/base/core/src/mgmt.h
@@ -3,7 +3,7 @@
 
 struct mgmt_t {
     size_t allocated;  // amount of bytes allocated (only takes tensors into account)
-    uint64_t ntensors; // number registered tensors (should take around 584942 years to overflow at 1Mio increments/s)
+    size_t ntensors; // number registered tensors (should take around 584942 years to overflow at 1Mio increments/s)
 } mgmt;
 
 void init_mgmt();
